+++ examples/search-in-graph/src/app/app.component.html
@@ -28,6 +28,9 @@
 </ng-template>
 <!-- end graph context menu template -->
 
+<!--<sz-entity-detail-graph-control></sz-entity-detail-graph-control>
+<sz-entity-detail-graph-filter></sz-entity-detail-graph-filter>-->
+
 <button (click)="onTabClick('detail')">show detail</button>
 <button (click)="onTabClick('filters')">show filters</button>
 <div class="content-container">
