--- src/lib/entity/detail/sz-entity-detail-graph/sz-entity-detail-graph-control.component.html
@@ -1,1 +1,0 @@
-<h2>SzEntityDetailGraphControlComponent works!</h2>
