--- src/lib/entity/detail/sz-entity-detail-graph/sz-entity-detail-graph-filter.component.html
@@ -1,1 +1,0 @@
-<h2>SzEntityDetailGraphFilterComponent works!</h2>
