+++ src/lib/entity/detail/sz-entity-detail-graph/sz-entity-detail-graph.component.old.html
@@ -1,5 +1,20 @@
 
-<div #graphContainer class="sz-standalone-graph-container">
+<div data-html2canvas-ignore class="section-header__wrapper graph-relationships"
+  (click)="toggleExpanded($event)"
+>
+  <div class="section-header__left-content">
+    <svg class="arrow" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
+      viewBox="0 0 386.257 386.257" width="12" xml:space="preserve">
+      <polygon points="0,96.879 193.129,289.379 386.257,96.879 "/></svg>
+    <svg xmlns="http://www.w3.org/2000/svg" class="mat-icon" width="24" height="24" viewBox="0 0 24 24"><circle cx="7.2" cy="14.4" r="3.2"/><circle cx="14.8" cy="18" r="2"/><circle cx="15.2" cy="8.8" r="4.8"/></svg>
+    <span class="section-header__title">{{ this.title }}</span>
+  </div>
+  <div class="section-header__right-content">
+    <span class="section-header__count">{{ data?.relatedEntities?.length }}</span>
+    <span class="section-header__countLabel">Relationships</span>
+  </div>
+</div>
+<div #graphContainer class="sz-entity-detail-section-collapsible-card">
   <div *ngIf="showZoomControl" class="zoom-control-container"
   [class.top-left]="zoomControlPosition == 'top-left'"
   [class.top-right]="zoomControlPosition == 'top-right'"
@@ -10,7 +25,6 @@
     <span><button class="zoom-in" (click)="zoomIn()"></button></span>
   </div>
   <sz-relationship-network #graph class="sz-relationship-network-graph"
-    [class.filters-showing]="this.showFiltersControl"
     svgViewBox="150 50 400 300"
     svgPreserveAspectRatio="xMinYMid meet"
     [entityIds]="graphIds"
@@ -18,9 +32,10 @@
     [buildOut]="buildOut"
     [linkGravity]="5"
     [highlight]="entityNodeColors"
-    [filter]="entityNodeFilters"
-    [includes]="entityMatchFilter"
+    [filter]="entityNodeFilterByDataSource"
     [showLinkLabels]="_showMatchKeys"
+    [captureMouseWheel]="captureMouseWheel"
+    (onWheelScroll)="onGraphScrollEvent($event)"
     (contextMenuClick)="onRightClick($event)"
     (entityClick)="onEntityClick($event)"
     (noResults)="onNoResults($event)"
@@ -34,15 +49,8 @@
     ></sz-graph-control>
 
     <sz-graph-filter *ngIf="showFiltersControl" class="sz-graph-filter"
-      [ngStyle]="{'width.px': filterWidth}"
-      [class.top-left]="filterControlPosition == 'top-left'"
-      [class.top-right]="filterControlPosition == 'top-right'"
-      [class.bottom-right]="filterControlPosition == 'bottom-right'"
-      [class.bottom-left]="filterControlPosition == 'bottom-left'"
-      [showLinkLabels]="_showMatchKeys"
-      (optionChanged)="onOptionChange($event)"
-      [showDataSources]="filterShowDataSources"
-      [showMatchKeys]="filterShowMatchKeys"
+    [showLinkLabels]="_showMatchKeys"
+    (optionChanged)="onOptionChange($event)"
     ></sz-graph-filter>
 
     <svg class="popout-icon"
@@ -60,6 +68,3 @@
       </g>
     </svg>
 </div>
-<!--
-  filterControlPosition: "{{ filterControlPosition }}"<br/>
-  entityIds: {{ graphIds }} -->
